+++ layers/Schema/packages/menus/src/FieldResolvers/MenuFieldResolver.php
@@ -5,16 +5,46 @@ declare(strict_types=1);
 namespace PoPSchema\Menus\FieldResolvers;
 
 use PoP\ComponentModel\FieldResolvers\AbstractDBDataFieldResolver;
+use PoP\ComponentModel\HelperServices\SemverHelperServiceInterface;
+use PoP\ComponentModel\Instances\InstanceManagerInterface;
 use PoP\ComponentModel\Misc\GeneralUtils;
+use PoP\ComponentModel\Schema\FieldQueryInterpreterInterface;
 use PoP\ComponentModel\Schema\SchemaDefinition;
 use PoP\ComponentModel\Schema\SchemaTypeModifiers;
 use PoP\ComponentModel\TypeResolvers\TypeResolverInterface;
+use PoP\Engine\CMS\CMSServiceInterface;
+use PoP\Hooks\HooksAPIInterface;
+use PoP\LooseContracts\NameResolverInterface;
+use PoP\Translation\TranslationAPIInterface;
 use PoPSchema\Menus\Facades\MenuTypeAPIFacade;
+use PoPSchema\Menus\ObjectModels\MenuItem;
+use PoPSchema\Menus\RuntimeRegistries\MenuItemRuntimeRegistryInterface;
 use PoPSchema\Menus\TypeResolvers\MenuItemTypeResolver;
 use PoPSchema\Menus\TypeResolvers\MenuTypeResolver;
 
 class MenuFieldResolver extends AbstractDBDataFieldResolver
 {
+    public function __construct(
+        TranslationAPIInterface $translationAPI,
+        HooksAPIInterface $hooksAPI,
+        InstanceManagerInterface $instanceManager,
+        FieldQueryInterpreterInterface $fieldQueryInterpreter,
+        NameResolverInterface $nameResolver,
+        CMSServiceInterface $cmsService,
+        SemverHelperServiceInterface $semverHelperService,
+        protected MenuItemRuntimeRegistryInterface $menuItemRuntimeRegistry,
+    ) {
+        parent::__construct(
+            $translationAPI,
+            $hooksAPI,
+            $instanceManager,
+            $fieldQueryInterpreter,
+            $nameResolver,
+            $cmsService,
+            $semverHelperService,
+        );
+    }
+
     public function getClassesToAttachTo(): array
     {
         return array(MenuTypeResolver::class);
@@ -23,6 +53,7 @@ class MenuFieldResolver extends AbstractDBDataFieldResolver
     public function getFieldNamesToResolve(): array
     {
         return [
+            'items',
             'itemDataEntries',
         ];
     }
@@ -30,7 +61,7 @@ class MenuFieldResolver extends AbstractDBDataFieldResolver
     public function getSchemaFieldType(TypeResolverInterface $typeResolver, string $fieldName): string
     {
         $types = [
-            // 'items' => SchemaDefinition::TYPE_ID,
+            'items' => SchemaDefinition::TYPE_ID,
             'itemDataEntries' => SchemaDefinition::TYPE_OBJECT,
         ];
         return $types[$fieldName] ?? parent::getSchemaFieldType($typeResolver, $fieldName);
@@ -39,7 +70,9 @@ class MenuFieldResolver extends AbstractDBDataFieldResolver
     public function getSchemaFieldTypeModifiers(TypeResolverInterface $typeResolver, string $fieldName): ?int
     {
         return match ($fieldName) {
-            'itemDataEntries' => SchemaTypeModifiers::NON_NULLABLE | SchemaTypeModifiers::IS_ARRAY,
+            'items',
+            'itemDataEntries'
+                => SchemaTypeModifiers::NON_NULLABLE | SchemaTypeModifiers::IS_ARRAY,
             default => parent::getSchemaFieldTypeModifiers($typeResolver, $fieldName),
         };
     }
@@ -62,7 +95,8 @@ class MenuFieldResolver extends AbstractDBDataFieldResolver
     public function getSchemaFieldDescription(TypeResolverInterface $typeResolver, string $fieldName): ?string
     {
         $descriptions = [
-            'itemDataEntries' => $this->translationAPI->__('The menu items', 'menus'),
+            'items' => $this->translationAPI->__('The menu items', 'menus'),
+            'itemDataEntries' => $this->translationAPI->__('The data for the menu items', 'menus'),
         ];
         return $descriptions[$fieldName] ?? parent::getSchemaFieldDescription($typeResolver, $fieldName);
     }
@@ -91,7 +125,7 @@ class MenuFieldResolver extends AbstractDBDataFieldResolver
                 $entries = array();
                 if ($itemsData) {
                     // Load these item data-fields. If other set needed, create another $field
-                    $item_data_fields = array('id', 'title', 'alt', 'classes', 'url', 'target', 'parentID', 'objectID', 'additionalAttrs');
+                    $item_data_fields = array('id', 'title', 'alt', 'classes', 'url', 'target', 'parentID', 'objectID', 'description');
                     /**
                      * @var MenuItemTypeResolver
                      */
@@ -138,6 +172,44 @@ class MenuFieldResolver extends AbstractDBDataFieldResolver
                     $arrangedEntriesPointer[] = $menuItemData;
                 }
                 return $arrangedEntries;
+            case 'items':
+                // Fetch the data for the items, and from it build the objects respecting their position in the menu
+                $itemDataEntries = $typeResolver->resolveValue(
+                    $resultItem,
+                    $this->fieldQueryInterpreter->getField(
+                        'itemDataEntries',
+                        [
+                            'flat' => true,
+                        ]
+                    ),
+                    $variables,
+                    $expressions,
+                    $options
+                );
+                if (GeneralUtils::isError($itemDataEntries)) {
+                    return $itemDataEntries;
+                }
+                // Build the MenuItem objects from the data, and save them on the dynamic registry
+                $topLevelMenuItemIDs = [];
+                foreach ($itemDataEntries as $menuItemData) {
+                    // Top-level items are those with no parent
+                    if ($menuItemData['parentID'] === null) {
+                        $topLevelMenuItemIDs[] = $menuItemData['id'];
+                    }
+                    $this->menuItemRuntimeRegistry->storeMenuItem(new MenuItem(
+                        $menuItemData['id'],
+                        $menuItemData['objectID'],
+                        $menuItemData['parentID'],
+                        $menuItemData['title'],
+                        $menuItemData['url'],
+                        $menuItemData['description'],
+                        $menuItemData['classes'],
+                        $menuItemData['target'],
+                    ));
+                }
+
+                // Return the IDs for the top-level items
+                return $topLevelMenuItemIDs;
         }
 
         return parent::resolveValue($typeResolver, $resultItem, $fieldName, $fieldArgs, $variables, $expressions, $options);
@@ -159,13 +231,13 @@ class MenuFieldResolver extends AbstractDBDataFieldResolver
         return 0;
     }
 
-    // public function resolveFieldTypeResolverClass(TypeResolverInterface $typeResolver, string $fieldName): ?string
-    // {
-    //     switch ($fieldName) {
-    //         case 'items':
-    //             return MenuItemTypeResolver::class;
-    //     }
+    public function resolveFieldTypeResolverClass(TypeResolverInterface $typeResolver, string $fieldName): ?string
+    {
+        switch ($fieldName) {
+            case 'items':
+                return MenuItemTypeResolver::class;
+        }
 
-    //     return parent::resolveFieldTypeResolverClass($typeResolver, $fieldName);
-    // }
+        return parent::resolveFieldTypeResolverClass($typeResolver, $fieldName);
+    }
 }