+++ layers/Schema/packages/menus/src/FieldResolvers/MenuItemFieldResolver.php
@@ -4,14 +4,44 @@ declare(strict_types=1);
 
 namespace PoPSchema\Menus\FieldResolvers;
 
-use PoP\ComponentModel\Schema\SchemaDefinition;
 use PoP\ComponentModel\FieldResolvers\AbstractDBDataFieldResolver;
+use PoP\ComponentModel\HelperServices\SemverHelperServiceInterface;
+use PoP\ComponentModel\Instances\InstanceManagerInterface;
+use PoP\ComponentModel\Schema\FieldQueryInterpreterInterface;
+use PoP\ComponentModel\Schema\SchemaDefinition;
+use PoP\ComponentModel\Schema\SchemaTypeModifiers;
 use PoP\ComponentModel\TypeResolvers\TypeResolverInterface;
-use PoPSchema\Menus\TypeResolvers\MenuItemTypeResolver;
+use PoP\Engine\CMS\CMSServiceInterface;
+use PoP\Hooks\HooksAPIInterface;
+use PoP\LooseContracts\NameResolverInterface;
+use PoP\Translation\TranslationAPIInterface;
 use PoPSchema\Menus\Facades\MenuItemTypeAPIFacade;
+use PoPSchema\Menus\RuntimeRegistries\MenuItemRuntimeRegistryInterface;
+use PoPSchema\Menus\TypeResolvers\MenuItemTypeResolver;
 
 class MenuItemFieldResolver extends AbstractDBDataFieldResolver
 {
+    public function __construct(
+        TranslationAPIInterface $translationAPI,
+        HooksAPIInterface $hooksAPI,
+        InstanceManagerInterface $instanceManager,
+        FieldQueryInterpreterInterface $fieldQueryInterpreter,
+        NameResolverInterface $nameResolver,
+        CMSServiceInterface $cmsService,
+        SemverHelperServiceInterface $semverHelperService,
+        protected MenuItemRuntimeRegistryInterface $menuItemRuntimeRegistry,
+    ) {
+        parent::__construct(
+            $translationAPI,
+            $hooksAPI,
+            $instanceManager,
+            $fieldQueryInterpreter,
+            $nameResolver,
+            $cmsService,
+            $semverHelperService,
+        );
+    }
+
     public function getClassesToAttachTo(): array
     {
         return array(MenuItemTypeResolver::class);
@@ -20,12 +50,15 @@ class MenuItemFieldResolver extends AbstractDBDataFieldResolver
     public function getFieldNamesToResolve(): array
     {
         return [
+            // This field is special in that it is retrieved from the registry
+            'children',
+            // All other fields are properties in the object
             'title',
             'alt',
             'url',
             'classes',
             'target',
-            'additionalAttrs',
+            'description',
             'objectID',
             'parentID',
         ];
@@ -34,27 +67,39 @@ class MenuItemFieldResolver extends AbstractDBDataFieldResolver
     public function getSchemaFieldType(TypeResolverInterface $typeResolver, string $fieldName): string
     {
         $types = [
+            'children' => SchemaDefinition::TYPE_ID,
             'title' => SchemaDefinition::TYPE_STRING,
             'alt' => SchemaDefinition::TYPE_STRING,
             'url' => SchemaDefinition::TYPE_URL,
             'classes' => SchemaDefinition::TYPE_STRING,
             'target' => SchemaDefinition::TYPE_STRING,
-            'additionalAttrs' => SchemaDefinition::TYPE_STRING,
+            'description' => SchemaDefinition::TYPE_STRING,
             'objectID' => SchemaDefinition::TYPE_ID,
             'parentID' => SchemaDefinition::TYPE_ID,
         ];
         return $types[$fieldName] ?? parent::getSchemaFieldType($typeResolver, $fieldName);
     }
 
+    public function getSchemaFieldTypeModifiers(TypeResolverInterface $typeResolver, string $fieldName): ?int
+    {
+        return match ($fieldName) {
+            'children',
+            'classes'
+                => SchemaTypeModifiers::NON_NULLABLE | SchemaTypeModifiers::IS_ARRAY,
+            default => parent::getSchemaFieldTypeModifiers($typeResolver, $fieldName),
+        };
+    }
+
     public function getSchemaFieldDescription(TypeResolverInterface $typeResolver, string $fieldName): ?string
     {
         $descriptions = [
+            'children' => $this->translationAPI->__('Menu item children items', 'menus'),
             'title' => $this->translationAPI->__('Menu item title', 'menus'),
             'alt' => $this->translationAPI->__('Menu item alt', 'menus'),
             'url' => $this->translationAPI->__('Menu item URL', 'menus'),
             'classes' => $this->translationAPI->__('Menu item classes', 'menus'),
             'target' => $this->translationAPI->__('Menu item target', 'menus'),
-            'additionalAttrs' => $this->translationAPI->__('Menu item additional attributes', 'menus'),
+            'description' => $this->translationAPI->__('Menu item additional attributes', 'menus'),
             'objectID' => $this->translationAPI->__('ID of the object linked to by the menu item ', 'menus'),
             'parentID' => $this->translationAPI->__('Menu item\'s parent ID', 'menus'),
         ];
@@ -79,6 +124,8 @@ class MenuItemFieldResolver extends AbstractDBDataFieldResolver
         $menuItemTypeAPI = MenuItemTypeAPIFacade::getInstance();
         $menuItem = $resultItem;
         switch ($fieldName) {
+            case 'children':
+                return array_keys($this->menuItemRuntimeRegistry->getMenuItemChildren($typeResolver->getID($menuItem)));
             case 'title':
             case 'alt':
                 return $menuItemTypeAPI->getMenuItemTitle($menuItem);
@@ -87,26 +134,16 @@ class MenuItemFieldResolver extends AbstractDBDataFieldResolver
                 return $menuItemTypeAPI->getMenuItemURL($menuItem);
 
             case 'classes':
-                // Copied from nav-menu-template.php function start_el
-                $classes = $menuItemTypeAPI->getMenuItemClasses($menuItem);
-                $classes = empty($classes) ? array() : (array) $classes;
-                $classes[] = 'menu-item';
-                $classes[] = 'menu-item-' . $menuItemTypeAPI->getMenuItemID($menuItem);
-                if ($parentID = $menuItemTypeAPI->getMenuItemParentID($menuItem)) {
-                    $classes[] = 'menu-item-parent';
-                    $classes[] = 'menu-item-parent-' . $parentID;
-                }
-                if ($objectID = $menuItemTypeAPI->getMenuItemObjectID($menuItem)) {
-                    $classes[] = 'menu-item-object-id-' . $objectID;
-                }
-                return join(' ', $this->hooksAPI->applyFilters('menuitem:classes', array_filter($classes), $menuItem, array()));
+                return $this->hooksAPI->applyFilters(
+                    'menuitem:classes',
+                    $menuItemTypeAPI->getMenuItemClasses($menuItem),
+                    $menuItem
+                );
 
             case 'target':
                 return $menuItemTypeAPI->getMenuItemTarget($menuItem);
 
-            case 'additionalAttrs':
-                // Using the description, because WP does not give a field for extra attributes when creating a menu,
-                // and this is needed to add target="addons" for the Add ContentPost link
+            case 'description':
                 return $menuItemTypeAPI->getMenuItemDescription($menuItem);
 
             case 'objectID':
@@ -117,5 +154,15 @@ class MenuItemFieldResolver extends AbstractDBDataFieldResolver
         }
 
         return parent::resolveValue($typeResolver, $resultItem, $fieldName, $fieldArgs, $variables, $expressions, $options);
+    }
+
+    public function resolveFieldTypeResolverClass(TypeResolverInterface $typeResolver, string $fieldName): ?string
+    {
+        switch ($fieldName) {
+            case 'children':
+                return MenuItemTypeResolver::class;
+        }
+
+        return parent::resolveFieldTypeResolverClass($typeResolver, $fieldName);
     }
 }