+++ layers/Schema/packages/menus/src/TypeDataLoaders/MenuItemTypeDataLoader.php
@@ -4,16 +4,32 @@ declare(strict_types=1);
 
 namespace PoPSchema\Menus\TypeDataLoaders;
 
+use PoP\ComponentModel\Instances\InstanceManagerInterface;
 use PoP\ComponentModel\TypeDataLoaders\AbstractTypeDataLoader;
-use PoPSchema\Menus\Facades\MenuItemTypeAPIFacade;
+use PoP\Hooks\HooksAPIInterface;
+use PoP\LooseContracts\NameResolverInterface;
+use PoPSchema\Menus\RuntimeRegistries\MenuItemRuntimeRegistryInterface;
 
 class MenuItemTypeDataLoader extends AbstractTypeDataLoader
 {
+    public function __construct(
+        HooksAPIInterface $hooksAPI,
+        InstanceManagerInterface $instanceManager,
+        NameResolverInterface $nameResolver,
+        protected MenuItemRuntimeRegistryInterface $menuItemRuntimeRegistry,
+    ) {
+        parent::__construct(
+            $hooksAPI,
+            $instanceManager,
+            $nameResolver,
+        );
+    }
+
     public function getObjects(array $ids): array
     {
-        $menuItemTypeAPI = MenuItemTypeAPIFacade::getInstance();
+        // Retrieve each item from the dynamic registry
         return array_map(
-            fn (string | int $id) => $menuItemTypeAPI->getMenuItem($id),
+            fn (string | int $id) => $this->menuItemRuntimeRegistry->getMenuItem($id),
             $ids
         );
     }